Revise the code for ipfs c api
