+++ c_api.h
@@ -3,9 +3,7 @@
 
 #include "c_secure_comm.h"
 
-#define IV_SIZE 16
 #define MAX 1000000
-#define BUFF_SIZE 100
 
 // Load config file from path and save the information in ctx struct.
 // Also loads public and private key in EVP_PKEY struct.
@@ -80,25 +78,5 @@ void free_session_key_list_t(session_key_list_t *session_key_list);
 // Free memory used in SST_ctx recursively.
 // @param SST_ctx_t loaded SST_ctx_t to free
 void free_SST_ctx_t(SST_ctx_t *ctx);
-
-// Do command "ipfs add command" and save the hash value.
-void ipfs_add_command_save_result();
-
-// Encrypt the file with sessionkey and upload the file in IPFS environment.
-// @param SST_session_ctx_t session_ctx to encrypt the file
-void file_encrypt_upload(SST_session_ctx_t *session_ctx);
-
-// Download the file in IPFS environment and decrypt the file with sessionkey.
-// @param SST_session_ctx_t session_ctx to decrypt the file
-void file_download_decrypt(SST_session_ctx_t *session_ctx);
-
-// Request the data to datacenter
-// @param SST_session_ctx_t session_ctx SST_ctx_t ctx to upload the data to datacenter.
-void upload_to_datamanagement(SST_session_ctx_t *session_ctx, SST_ctx_t *ctx);
-
-// Receive the data from datacenter
-// @param SST_session_ctx_t session_ctx SST_ctx_t ctx to download the data from datacenter.
-void download_from_datamanagement(SST_session_ctx_t *session_ctx, SST_ctx_t *ctx);
-
 
 #endif  // C_API_H
