+++ examples/c_client.config
@@ -1,5 +1,5 @@
 entityInfo.name=net1.client
-entityInfo.purpose.ms={"group":"Servers"}
+entityInfo.purpose={"group":"Servers"}
 entityInfo.number_key=3
 authInfo.pubkey.path=../../../auth_certs/Auth101EntityCert.pem
 entityInfo.privkey.path=../../../credentials/keys/net1/Net1.ClientKey.pem
