+++ ipfs_examples/CMakeLists.txt
@@ -9,8 +9,8 @@ project(SST VERSION 1.0.0 LANGUAGES C)
 find_package(OpenSSL REQUIRED)
 find_package(Threads REQUIRED)
 
-add_executable(entity_client entity_client.c ../c_api.c ../c_common.c ../c_crypto.c ../c_secure_comm.c ../load_config.c)
-add_executable(entity_server entity_server.c ../c_api.c ../c_common.c ../c_crypto.c ../c_secure_comm.c ../load_config.c)
+add_executable(entity_client entity_client.c ../ipfs.c ../c_api.c ../c_common.c ../c_crypto.c ../c_secure_comm.c ../load_config.c)
+add_executable(entity_server entity_server.c ../ipfs.c ../c_api.c ../c_common.c ../c_crypto.c ../c_secure_comm.c ../load_config.c)
 
 target_link_libraries(entity_client OpenSSL::SSL)
 target_link_libraries(entity_client Threads::Threads)
