+++ ipfs_examples/c_client.config
@@ -1,6 +1,6 @@
 entityInfo.name=net1.uploader
-entityInfo.purpose.ms={"group":"Servers"}
-entityInfo.purpose.fs={"FileSharing":"FileSharingTeam"}
+entityInfo.purpose={"group":"Servers"}
+entityInfo.purpose={"FileSharing":"FileSharingTeam"}
 entityInfo.number_key=3
 authInfo.pubkey.path=../../../auth_certs/Auth101EntityCert.pem
 entityInfo.privkey.path=../../../credentials/keys/net1/Net1.UploaderKey.pem
