+++ ipfs_examples/c_server.config
@@ -1,5 +1,5 @@
 entityInfo.name=net1.downloader
-entityInfo.purpose.ms={"keyId":00000000}
+entityInfo.purpose={"keyId":00000000}
 entityInfo.number_key=1
 authInfo.pubkey.path=../../../auth_certs/Auth101EntityCert.pem
 entityInfo.privkey.path=../../../credentials/keys/net1/Net1.DownloaderKey.pem
