+++ ipfs_examples/entity_client.c
@@ -1,4 +1,4 @@
-#include "../c_api.h"
+#include "../ipfs.h"
 
 int main(int argc, char *argv[]) {
     char *config_path = argv[1];
